base_files: support raw LED registration in UCI
As well as all the existing UCI registration for LEDs, support simply
listing the name<->sysfs mapping.  The existing "ucidef_set_led_default"
changes the status of the led explicitly in /etc/init.d/leds which may
not be desirable.  Simply registering the name lets the /sbin/led.sh
utility work, but doesn't interfere with the LED itself.

Signed-off-by: Karl Palsson <karlp@remake.is>